backend/requirements.txt
Flask
flasgger
mysql-connector-python
flask-cors